requirements.txt
discord.py==2.7.1
openai==2.26.0
openai-agents==0.10.5
tinydb==4.8.2
python-dotenv==1.2.2